refactor(models): change DEFAULT_TIMEOUT type from u64 to u16
- More appropriate type for timeout values (max 65535 seconds)
- Add explicit casts where u64 is needed:
  - Dashboard timestamp arithmetic
  - tokio Duration::from_secs()
  - Session timeout checks

+++ src/models/dashboard.rs
@@ -290,7 +290,7 @@ impl Group {
                         if let Some(ready_at) = session.ready_at {
                             if let Ok(duration_since_epoch) = ready_at.duration_since(SystemTime::UNIX_EPOCH) {
                                 let ready_timestamp = duration_since_epoch.as_secs();
-                                let deadline_timestamp = ready_timestamp + DEFAULT_TIMEOUT;
+                                let deadline_timestamp = ready_timestamp + DEFAULT_TIMEOUT as u64;
                                 description.push_str(&format!("Join deadline: <t:{}:R>\n", deadline_timestamp));
                                 description.push_str("Missing players will be removed. Overflow players will take their spots.\n\n");
                             }
@@ -332,7 +332,7 @@ impl Group {
                         if let Some(ready_at) = current_session.ready_at {
                             if let Ok(duration_since_epoch) = ready_at.duration_since(SystemTime::UNIX_EPOCH) {
                                 let ready_timestamp = duration_since_epoch.as_secs();
-                                let deadline_timestamp = ready_timestamp + DEFAULT_TIMEOUT;
+                                let deadline_timestamp = ready_timestamp + DEFAULT_TIMEOUT as u64;
                                 description.push_str(&format!("Join deadline: <t:{}:R>\n", deadline_timestamp));
                                 description.push_str("Missing players will be removed. Overflow players will take their spots.\n\n");
                             }
